NewsSpider/requirements.txt
APScheduler==3.6.1
asn1crypto==0.24.0
attrs==19.1.0
Automat==0.7.0
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
constantly==15.1.0
cryptography==2.7
cssselect==1.1.0
Django==2.2.5
djangorestframework==3.10.3
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
lxml==4.4.1
parsel==1.5.2
pyasn1==0.4.7
pyasn1-modules==0.2.6
pycparser==2.19
PyDispatcher==2.0.5
PyHamcrest==1.9.0
pyOpenSSL==19.0.0
python-scrapyd-api==2.1.2
pytz==2019.2
queuelib==1.5.0
requests==2.22.0
Scrapy==1.7.3
scrapy-djangoitem==1.1.1
scrapyd==1.2.1
scrapyd-client==1.1.0
service-identity==18.1.0
six==1.12.0
sqlparse==0.3.0
Twisted==19.7.0
tzlocal==2.0.0
urllib3==1.25.3
w3lib==1.21.0
zope.interface==4.6.0
